fixed implicit int -> double casts
git-svn-id: svn://svn.code.sf.net/p/genomeview/code/genomeview/trunk@401 6512a8d4-c03c-0410-a19d-ba0f7ff4f5ce

+++ net/sf/genomeview/gui/annotation/track/ShortReadTrack.java
@@ -302,12 +302,12 @@ public class ShortReadTrack extends Track {
 								for (int j = rf.start(); j <= rf.end(); j++) {
 									char readNt = rf.getNucleotide(j - rf.start() + 1);
 									char refNt = Character.toUpperCase(entry.sequence.getNucleotide(j));
-									int tx1 = Convert.translateGenomeToScreen(j, r, screenWidth);
-									int tx2 = Convert.translateGenomeToScreen(j + 1, r, screenWidth);
+									double tx1 = Convert.translateGenomeToScreen(j, r, screenWidth);
+									double tx2 = Convert.translateGenomeToScreen(j + 1, r, screenWidth);
 
 									if (readNt != refNt) {
 										g.setColor(Color.ORANGE);
-										g.fillRect(tx1, rec.y, tx2 - tx1, rec.height);
+										g.fillRect((int)tx1, rec.y,(int)( tx2 - tx1), rec.height);
 										if (model.getAnnotationLocationVisible().length() < 100) {
 											// g.setColor(cachedColors.get(i));
 											g.setColor(c);
